Change Glyph#getUnicode return value from Integer to int.
DEVSIX-791

Autoported commit.
Original commit hash: [0481a2d0a0]

+++ itext/itext.io/itext/io/font/Type1Font.cs
@@ -131,7 +131,7 @@ namespace iText.IO.Font {
 
         public override int GetKerning(Glyph first, Glyph second) {
             if (first.HasValidUnicode() && second.HasValidUnicode()) {
-                long record = ((long)first.GetUnicode() << 32) + (int)second.GetUnicode();
+                long record = ((long)first.GetUnicode() << 32) + second.GetUnicode();
                 if (kernPairs.ContainsKey(record)) {
                     return (int)kernPairs.Get(record);
                 }
